requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
requests>=2.28.1
python-dotenv>=1.0.0
psycopg2-binary>=2.9.9
pytest>=7.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
httpx>=0.24.0
websockets>=10.0
langchain>=0.1.0
langchain-huggingface>=0.0.1
langchain-google-genai>=0.0.1
google-generativeai>=0.3.0
chromadb>=0.4.0
bitsandbytes>=0.41.0
accelerate>=0.26.0
mcp>=0.1.0
gradio>=3.0.0
PyYAML>=6.0.1
